+++ packages/wallet/src/blueprints/wallet.ts
@@ -7,12 +7,12 @@ import { setWalletBackupContent } from "../operations/event/wallet.js";
 import { setMintTags, setPrivateKeyTag } from "../operations/tag/wallet.js";
 
 /** A blueprint to create a new 17375 wallet */
-export function WalletBlueprint(privateKey: Uint8Array, mints: string[]): EventBlueprint {
+export function WalletBlueprint(mints: string[], privateKey?: Uint8Array): EventBlueprint {
   return (ctx) =>
     EventFactory.runProcess(
       { kind: WALLET_KIND },
       ctx,
-      modifyHiddenTags(setPrivateKeyTag(privateKey), setMintTags(mints)),
+      modifyHiddenTags(privateKey ? setPrivateKeyTag(privateKey) : undefined, setMintTags(mints)),
     );
 }